CU-86b2czk01 Adding unit test for realtabformer synthesizer
